HellerSet works with a MUCH stronger (false?) spec for myNext
If myNext carries the refinement of the set inside the reference (s in RGRef<p,r,g> (Set <s> a)) out
--- that the lower bound on the value the reference points to is given by the slack projection of
the argument node --- then the other goals discharge naturally.

But giving the natural spec, where the slack condition holds for nodes and delnodes but the
parameter is otherwise just true, isn't satisfied by the implementation:
    {\x -> (isNull l || isHead l || slack l < x)}

Need a working, operational myNext.